// src/types.ts
export type PokemonType =
  | 'fire'
  | 'water'
  | 'grass'
  | 'electric'
  | 'psychic'
  | 'ice'
  | 'dragon'
  | 'dark'
  | 'fairy'
  | 'normal'
  | 'fighting'
  | 'flying'
  | 'poison'
  | 'ground'
  | 'rock'
  | 'bug'
  | 'ghost'
  | 'steel';

export type TaskDifficulty = 'facil' | 'media' | 'dificil' | 'legendaria';

export type TaskCategory =
  | 'hogar_cocina'
  | 'hogar_fregar'
  | 'hogar_barrer'
  | 'hogar_ordenar'
  | 'hogar_basura'
  | 'hogar_jardin'
  | 'hogar_cristales'
  | 'estudio_general'
  | 'salud_ejercicio'
  | 'personal_finanzas'
  | 'personal_creatividad';